Fix new model hashes not being stored in metadata (#103)
Previously when a model's metadata is being saved to disk, we would check for a missing hash and generate it. We would save that generated hash into the file's metadata header, but not into the in-memory model or the model metadata cache. This resulted in the file's generated hash not being used cache was invalidated and the model's header was read again. The next time we would try to save the metadata for that model, we would detect the file hash was in the file's header, but once again not check if it was in-memory.

Now instead, we directly save the generated hash into our in-memory model, immediately save that into our metadata cache, and then finally we write it into the file's header with the other values.  When determining when we should generate a hash, we now look at both our in-memory model and the file header version.